softart:   demo changed (add a planar).   bug fix: error when an invisible triangle clipped.

+++ softart/softart/src/rasterizer.cpp
@@ -481,15 +481,12 @@ void rasterizer::rasterize_triangle(const vs_output& v0, const vs_output& v1, co
 	//±ß½çÌÞ³ý
 	
 	//±³ÃæÌÞ³ý
-	if(cm_ != cull_none)
-	{
+	if(cm_ != cull_none) {
 		float area = compute_area(v0, v1, v2);
-		if( (cm_ == cull_front) && (area > 0) )
-		{
+		if( (cm_ == cull_front) && (area > 0) ) {
 			return;
 		}
-		if( (cm_ == cull_back) && (area < 0) )
-		{
+		if( (cm_ == cull_back) && (area < 0) ) {
 			return;
 		}
 	}
@@ -503,14 +500,14 @@ void rasterizer::rasterize_triangle(const vs_output& v0, const vs_output& v1, co
 	} else
 	{
 		vector<vs_output> clipped_verts;
+
 		h_clipper clipper = pparent_->get_clipper();
 		clipper->clip(clipped_verts , pparent_->get_viewport() , v0, v1, v2);
 
-		for(size_t i_tri = 1; i_tri < clipped_verts.size() - 1; ++i_tri)
+		for(int i_tri = 1; i_tri < int(clipped_verts.size()) - 1; ++i_tri)
 		{
 			rasterize_triangle_impl(clipped_verts[0], clipped_verts[i_tri], clipped_verts[i_tri+1], vp);
 		}
-		//rasterize_triangle_impl(v0, v1, v2);
 	}
 }
